Refactor DrillCursor
Refactor DrillCursor to be more self-contained.
